+++ src/dialogs/MainWindow.h
@@ -202,6 +202,8 @@ private:
     void openFileList(const QStringList &fileNames);
     bool checkEditorsBeforeClose(const QVector<ScintillaNext *> &editors);
     bool checkFileForModification(ScintillaNext *editor);
+    bool resolveExternalSaveConflict(ScintillaNext *editor);
+    void openExternalVersion(ScintillaNext *editor);
     void showSaveErrorMessage(ScintillaNext *editor, QFileDevice::FileError error);
     void showEditorZoomLevelIndicator();
     void setupThemeMenu();